pyproject.toml
[tool.poetry]
name = "src"
version = "2022.06.06"
description = ""
authors = ["darrida <darrida.py@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
prefect = "^2.0.0"
# boto3 = "^1.24.3"
pandas = "^1.5.0"
psycopg2-binary = "^2.9.3"
prefect-sqlalchemy = "^0.2.0"
prefect-aws = "^0.1.6"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
